Added metrics parser
